Move ra-core record files into dedicated directory
