+++ packages/ra-core/src/controller/index.ts
@@ -64,8 +64,6 @@ export * from './details';
 export * from './show';
 export * from './field';
 export * from './input';
-export * from './RecordContext';
+export * from './record';
 export * from './saveModifiers';
 export * from './useList';
-export * from './useRecordContext';
-export * from './WithRecord';
