+++ packages/ra-core/src/controller/record/index.ts
@@ -1,0 +1,3 @@
+export * from './RecordContext';
+export * from './useRecordContext';
+export * from './WithRecord';
